# docs/api_doc/swagger.json
{"openapi":"3.1.0","info":{"title":"Media Service","description":"Service for media management","contact":{"name":"Sumar+","email":"-"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8083","description":"Service for media management"}],"paths":{"/api/v1/media/savefile":{"post":{"tags":["file-controller"],"summary":"Upload a file and optionally verify its integrity with hashes","description":"This endpoint allows you to upload a file. Optionally, you can provide MD5 or SHA-256 hashes for integrity checks.","operationId":"saveFile","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"},"hashMd5":{"type":"string"},"hashSha256":{"type":"string"}},"required":["file"]}}}},"responses":{"200":{"description":"File uploaded successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UuidResponseDto"}}}},"400":{"description":"Bad Request due to invalid input (e.g., missing file, invalid hash format, or invalid storage path)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"Not Found: File not found for the provided hash","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"409":{"description":"Conflict: File integrity compromised (e.g., hash mismatch for MD5 or SHA-256)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error due to unexpected issues (e.g., file system error, MIME type detection failure, or database interaction error)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/api/v1/media/getFileBase64/{uuid}":{"get":{"tags":["file-controller"],"summary":"Retrieve file in Base64 by UUID","description":"Fetches file data using its UUID and returns it encoded in Base64 and metadata within a DTO.","operationId":"getFileBase64","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"File retrieved successfully as a byte array","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"File not found or not registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"409":{"description":"File integrity conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/api/v1/media/getFile/{uuid}":{"get":{"tags":["file-controller"],"summary":"Retrieve file by UUID","description":"Fetches the file as a byte array using its UUID and returns metadata in the headers.","operationId":"getFile","parameters":[{"name":"uuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"File retrieved successfully as a byte array","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"File not found or not registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"409":{"description":"File integrity conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"UuidResponseDto":{"properties":{"uuid":{"type":"string","format":"uuid"}}},"ErrorApi":{"properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}