# src/main/java/com/platform/media_service/dtos/UuidResponseDto.java
package com.platform.media_service.dtos;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

/**
 * Data Transfer Object (DTO) for representing a UUID response.
 */
@Data
@AllArgsConstructor
@NoArgsConstructor
public class UuidResponseDto {
    /**
     * The unique identifier (UUID) to be returned in the response.
     */
    @JsonProperty(namespace = "uuid")
    private UUID uuid;
}